Move download stuff into src folder
